test(indexer): rename test, move to more suitable file
