+++ integration_tests/tests/bridge.rs
@@ -557,37 +557,6 @@ fn create_zone_indexer_observer(
     ))
 }
 
-/// Test that the indexer status RPC reports caught-up with no stall after a clean run.
-///
-/// TODO: Integration-level park testing (publishing a bad block to force a stall) is a follow-up
-/// needing fault injection support in the test harness.
-#[test]
-async fn indexer_status_rpc_reports_caught_up_with_no_stall() -> anyhow::Result<()> {
-    use indexer_service_rpc::RpcClient as _;
-
-    let ctx = TestContext::new().await?;
-
-    let indexer_tip = wait_for_indexer_to_catch_up(&ctx).await?;
-
-    let status = ctx.indexer_client().get_status().await?;
-    assert_eq!(
-        status.state,
-        indexer_service_protocol::IndexerSyncState::CaughtUp,
-        "indexer should be caught up, got {status:?}"
-    );
-    assert!(
-        status.stall_reason.is_none(),
-        "indexer should have no stall reason after a clean run, got {status:?}"
-    );
-    assert_eq!(
-        status.indexed_block_id,
-        Some(indexer_tip),
-        "status indexed_block_id should equal the caught-up tip"
-    );
-
-    Ok(())
-}
-
 async fn wait_for_finalized_withdraw_op(
     observer: &ZoneIndexer<NodeHttpClient>,
     expected_amount: u64,
